pyproject.toml
[project]
name = "hud-trace-explorer"
version = "0.1.0"
description = "HUD environment for exploring and analyzing trace data using coding tools"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "hud-python",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "ruff==0.11.2"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
only-include = ["env.py"]

[tool.hatch.build.targets.sdist]
only-include = ["env.py", "pyproject.toml", "README.md"]

[tool.hud]
name = "trace-explorer"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "B", "I", "C4", "UP"]
ignore = ["E501"]
